pyproject.toml
[tool.poetry]
name = "rbf-clan-bot"
version = "0.1.0"
description = ""
authors = ["SerHappy <lokhmanovkirill@gmail.com>"]
readme = "README.md"
packages = [{ include = "rbf_clan_bot" }]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.7"
sqlalchemy = { extras = ["asyncio"], version = "^2.0.25" }
alembic = "^1.13.1"
asyncpg = "^0.29.0"
loguru = "^0.7.2"
uvloop = "^0.19.0"
pydantic-settings = "^2.2.1"
aiofiles = "^23.2.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.1"

[tool.ruff]
exclude = [".venv", "alembic"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D100",
    "D104",
    "FA102",
    "ANN101",
    "A002",
    "TD003",
    "FIX002",
    "TD002",
    "RUF001",
    "RUF002",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
